feat: enter new package name in GUI
