+++ api-editor/gui/src/common/DeleteAllAnnotations.tsx
@@ -1,0 +1,66 @@
+import {
+    AlertDialog,
+    AlertDialogBody,
+    AlertDialogContent,
+    AlertDialogFooter,
+    AlertDialogHeader,
+    AlertDialogOverlay,
+    Button,
+    Heading,
+    Text as ChakraText,
+    VStack,
+} from '@chakra-ui/react';
+import React, { useRef, useState } from 'react';
+import { useAppDispatch } from '../app/hooks';
+import { resetAnnotations } from '../features/annotations/annotationSlice';
+
+export const DeleteAllAnnotations = function () {
+    const dispatch = useAppDispatch();
+    const [isOpen, setIsOpen] = useState(false);
+    const cancelRef = useRef(null);
+
+    // Event handlers ----------------------------------------------------------
+
+    const handleConfirm = () => {
+        dispatch(resetAnnotations());
+        setIsOpen(false);
+    };
+    const handleCancel = () => setIsOpen(false);
+
+    // Render ------------------------------------------------------------------
+
+    return (
+        <>
+            <Button onClick={() => setIsOpen(true)}>Delete all annotations</Button>
+
+            <AlertDialog isOpen={isOpen} leastDestructiveRef={cancelRef} onClose={handleCancel}>
+                <AlertDialogOverlay>
+                    <AlertDialogContent>
+                        <AlertDialogHeader>
+                            <Heading>Delete all annotations</Heading>
+                        </AlertDialogHeader>
+
+                        <AlertDialogBody>
+                            <VStack alignItems="flexStart">
+                                <ChakraText>Are you sure? You can't undo this action afterwards.</ChakraText>
+                                <ChakraText>
+                                    Hint: Consider exporting your work first by clicking on the "Export" button in the
+                                    menu bar.
+                                </ChakraText>
+                            </VStack>
+                        </AlertDialogBody>
+
+                        <AlertDialogFooter>
+                            <Button ref={cancelRef} onClick={handleCancel}>
+                                Cancel
+                            </Button>
+                            <Button colorScheme="red" onClick={handleConfirm} ml={3}>
+                                Delete
+                            </Button>
+                        </AlertDialogFooter>
+                    </AlertDialogContent>
+                </AlertDialogOverlay>
+            </AlertDialog>
+        </>
+    );
+};
